+++ include/RiscvEmulatorExtensionZbs.h
@@ -48,7 +48,7 @@ static inline void RiscvEmulatorBCLRI(
     void *rd,
     const uint8_t rs1num __attribute__((unused)),
     const void *rs1,
-    const uint32_t shamt) {
+    const uint8_t shamt) {
 
     if (rdnum == 0) {
         return;
@@ -87,7 +87,7 @@ static inline void RiscvEmulatorBEXTI(
     void *rd,
     const uint8_t rs1num __attribute__((unused)),
     const void *rs1,
-    const uint32_t shamt) {
+    const uint8_t shamt) {
 
     if (rdnum == 0) {
         return;
@@ -126,7 +126,7 @@ static inline void RiscvEmulatorBINVI(
     void *rd,
     const uint8_t rs1num __attribute__((unused)),
     const void *rs1,
-    const uint32_t shamt) {
+    const uint8_t shamt) {
 
     if (rdnum == 0) {
         return;
@@ -165,7 +165,7 @@ static inline void RiscvEmulatorBSETI(
     void *rd,
     const uint8_t rs1num __attribute__((unused)),
     const void *rs1,
-    const uint32_t shamt) {
+    const uint8_t shamt) {
 
     if (rdnum == 0) {
         return;
